requirements.txt
streamlit
face_recognition
dlib==19.8.1
imutils
opencv-python==4.3.0.36
pillow
numpy
#pandas
